# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "C3M Centralia API",
    "version": "1.0.0",
    "description": "Business management, clinical records, and appointment booking system",
    "contact": {
      "name": "API Support",
      "email": "support@c3mcentral.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://c3m-centralia-backend.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "business": {
            "type": "string",
            "description": "Business ID"
          },
          "name": {
            "type": "string",
            "example": "Massage Therapy"
          },
          "durationMinutes": {
            "type": "number",
            "example": 60
          },
          "price": {
            "type": "number",
            "example": 75
          },
          "description": {
            "type": "string",
            "example": "Relaxing full body massage"
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/businesses/{businessId}/services": {
      "post": {
        "summary": "Create a new service for a business (Owner only)",
        "description": "Only the business owner can create services for their business",
        "tags": ["Services"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Business ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "duration", "price"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Massage Therapy"
                  },
                  "description": {
                    "type": "string",
                    "example": "Relaxing full body massage"
                  },
                  "duration": {
                    "type": "number",
                    "example": 60
                  },
                  "price": {
                    "type": "number",
                    "example": 75
                  },
                  "category": {
                    "type": "string",
                    "example": "Wellness"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "get": {
        "summary": "Get all services for a business",
        "tags": ["Services"],
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Business ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "services": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Service"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/businesses/{businessId}/services/{serviceId}": {
      "get": {
        "summary": "Get service by ID",
        "tags": ["Services"],
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Business ID"
          },
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "put": {
        "summary": "Update service (Owner only)",
        "description": "Only the business owner can update their services",
        "tags": ["Services"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Business ID"
          },
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "price": {
                    "type": "number"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "delete": {
        "summary": "Delete service (Owner only)",
        "description": "Only the business owner can delete their services",
        "tags": ["Services"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Business ID"
          },
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
...
  }
}